Renaming to quinine. Resolves #12.
