+++ README.md
@@ -1,3 +1,3 @@
-# kumQUAt: A suite of QUAlity control tools.
+# quinine: A refreshing treatment for all quality control ailments
 
 Read and variant metrics, useable for pipeline quality control purposes
