+++ quinine-core/src/test/resources/small.vcf
